refactor(api): split certificate v2/v3 in two files
